# Type specialization for faster kmer iteration