--- a/REQUIRE
+++ b/REQUIRE
@@ -7,4 +7,3 @@ Combinatorics
 IndexableBitVectors 1.0.0
 IntervalTrees
 Twiddle 1.0.0
-DataStructures