--- a/src/eachkmer.jl
+++ b/src/eachkmer.jl
@@ -6,32 +6,23 @@
 # This file is a part of BioJulia.
 # License is MIT: https://github.com/BioJulia/BioSequences.jl/blob/master/LICENSE.md
 
-# Iterate through every k-mer in a nucleotide sequence
-struct EachKmerIterator{T<:Kmer,S<:Sequence}
+abstract type AbstractKmerIterator{T,S} end
+
+struct EveryKmerIterator{T<:Kmer,S<:Sequence} <: AbstractKmerIterator{T,S}
     seq::S
-    step::Int
     start::Int
+    stop::Int
 end
 
-"""
-    each(::Type{Kmer{T,k}}, seq::Sequence[, step=1])
-
-Initialize an iterator over all k-mers in a sequence `seq` skipping ambiguous
-nucleotides without changing the reading frame.
-
-# Arguments
-* `Kmer{T,k}`: k-mer type to enumerate.
-* `seq`: a nucleotide sequence.
-* `step=1`: the number of positions between iterated k-mers
-
-# Examples
-```
-# iterate over DNA codons
-for (pos, codon) in each(DNAKmer{3}, dna"ATCCTANAGNTACT", 3)
-    @show pos, codon
+struct SpacedKmerIterator{T<:Kmer,S<:Sequence} <: AbstractKmerIterator{T,S}
+    seq::S
+    start::Int
+    step::Int
+    stop::Int
+    filled::Int # This is cached values for speed
+    increment::Int # This is cached values for speed
 end
-```
-"""
+
 function each(::Type{Kmer{T,K}}, seq::Sequence, step::Integer=1) where {T,K}
     if eltype(seq) ∉ (DNA, RNA)
         throw(ArgumentError("element type must be either DNA or RNA nucleotide"))
@@ -40,66 +31,134 @@ function each(::Type{Kmer{T,K}}, seq::Sequence, step::Integer=1) where {T,K}
     elseif step < 1
         throw(ArgumentError("step size must be positive"))
     end
-    return EachKmerIterator{Kmer{T,K},typeof(seq)}(seq, step, 1)
+    if step == 1
+        return EveryKmerIterator{Kmer{T,K},typeof(seq)}(seq, 1, lastindex(seq))
+    else
+        filled = max(0, K - step)
+        increment = max(1, step - K + 1)
+        return SpacedKmerIterator{Kmer{T,K},typeof(seq)}(seq, 1, step, lastindex(seq), filled, increment)
+    end
 end
 
 eachkmer(seq::BioSequence{A}, K::Integer, step::Integer=1) where {A<:DNAAlphabet} = each(DNAKmer{Int(K)}, seq, step)
 eachkmer(seq::BioSequence{A}, K::Integer, step::Integer=1) where {A<:RNAAlphabet} = each(RNAKmer{Int(K)}, seq, step)
 eachkmer(seq::ReferenceSequence, K::Integer, step::Integer=1) = each(DNAKmer{Int(K)}, seq, step)
 
-Base.eltype(::Type{EachKmerIterator{T,S}}) where {T,S} = Tuple{Int,T}
+Base.eltype(::Type{<:AbstractKmerIterator{T,S}}) where {T,S} = Tuple{Int,T}
+Base.IteratorSize(::Type{<:AbstractKmerIterator{T,S}}) where {T,S} = Base.SizeUnknown()
 
-function Base.IteratorSize(::Type{EachKmerIterator{T,S}}) where {T,S}
-    return Base.SizeUnknown()
-end
+const kmerbits = (0xff, 0x00, 0x01, 0xff,
+                  0x02, 0xff, 0xff, 0xff,
+                  0x03, 0xff, 0xff, 0xff,
+                  0xff, 0xff, 0xff, 0xff)
 
-Base.iterate(it::EachKmerIterator) = iterate(it, (it.start, UInt64(0)))
-
-function Base.iterate(
-        it::EachKmerIterator{T},
-        state::Tuple{Int, UInt64}) where {T}
-
-    k = kmersize(T)
-    pos, kmer = state
-    isok = true
-
-    # faster path: return the next overlapping kmer if possible
-    if it.step < k && pos + k - 1 ≤ lastindex(it.seq) && pos != it.start
-        offset = k - it.step
-        if it.step == 1
-            nt = inbounds_getindex(it.seq, pos+offset)
-            kmer = kmer << 2 | trailing_zeros(nt)
-            isok &= iscertain(nt)
-        else
-            for i in 1:it.step
-                nt = inbounds_getindex(it.seq, pos+i-1+offset)
-                kmer = kmer << 2 | trailing_zeros(nt)
-                isok &= iscertain(nt)
-            end
+function Base.iterate(it::AbstractKmerIterator{T,S}) where {T,S<:BioSequence{<:TwoBitNucs}}
+    filled, i, kmer = 0, 1, UInt64(0)
+    while i ≤ it.stop
+        nt = reinterpret(Int8, inbounds_getindex(it.seq, i))
+        @inbounds val = kmerbits[nt + 1]
+        kmer = kmer << 2 | val
+        filled += 1
+        if filled == kmersize(T)
+            return (1, T(kmer)), (i, kmer)
         end
-        if isok
-            return (pos, T(kmer)), (pos+it.step, kmer)
-        end
+        i += 1
     end
-
-    while pos + k - 1 ≤ lastindex(it.seq)
-        kmer, ok = extract_kmer_impl(it.seq, pos, k)
-        if ok
-            return (pos, T(kmer)), (pos+it.step, kmer)
-        end
-        pos += it.step
-    end
-
     return nothing
 end
 
-function extract_kmer_impl(seq, from, k)
-    kmer::UInt64 = 0
-    isok = true
-    for i in 1:k
-        nt = inbounds_getindex(seq, from+i-1)
-        kmer = kmer << 2 | trailing_zeros(nt)
-        isok &= iscertain(nt)
+function Base.iterate(it::EveryKmerIterator{T,S}, state::Tuple{Int,UInt64}) where {T,S<:BioSequence{<:TwoBitNucs}}
+    i, kmer = state
+    i += 1
+    pos = i - kmersize(T) + 1
+    if i > it.stop
+        return nothing
+    else
+        nt = reinterpret(Int8, inbounds_getindex(it.seq, i))
+        @inbounds val = kmerbits[nt + 1]
+        kmer = kmer << 2 | val
+        return (pos, T(kmer)), (i, kmer)
     end
-    return kmer, isok
+end
+
+function Base.iterate(it::SpacedKmerIterator{T,S}, state::Tuple{Int,UInt64}) where {T,S<:BioSequence{<:TwoBitNucs}}
+    i, kmer = state
+    filled = it.filled
+    i += max(1, it.step - kmersize(T) + 1)
+    while i ≤ it.stop
+        nt = reinterpret(Int8, inbounds_getindex(it.seq, i))
+        val = kmerbits[nt + 1] # inbounds
+        kmer = kmer << 2 | val
+        filled += 1
+        if filled == kmersize(T)
+            pos = i - kmersize(T) + 1
+            return (pos, T(kmer)), (i, kmer)
+        end
+        i += 1
+    end
+    return nothing
+end
+
+function Base.iterate(it::AbstractKmerIterator{T,S}) where {T,S<:BioSequence{<:FourBitNucs}}
+    filled, i, kmer = 0, 1, UInt64(0)
+    while i ≤ it.stop
+        nt = reinterpret(Int8, inbounds_getindex(it.seq, i))
+        if nt == 0xff # ambiguous
+            filled = 0
+        else
+            filled += 1
+            @inbounds val = kmerbits[nt + 1]
+            kmer = kmer << 2 | val
+        end
+        if filled == kmersize(T)
+            pos = i - kmersize(T) + 1
+            return (pos, T(kmer)), (i, kmer)
+        end
+        i += 1
+    end
+    return nothing
+end
+
+function Base.iterate(it::EveryKmerIterator{T,S}, state::Tuple{Int,UInt64}) where {T,S<:BioSequence{<:FourBitNucs}}
+    i, kmer = state
+    i += 1
+    filled = kmersize(T) - 1
+    while i ≤ it.stop
+        nt = reinterpret(Int8, inbounds_getindex(it.seq, i))
+        if nt == 0xff # ambiguous
+            filled = 0
+        else
+            filled += 1
+            @inbounds val = kmerbits[nt + 1]
+            kmer = kmer << 2 | val
+        end
+        if filled == kmersize(T)
+            pos = i - kmersize(T) + 1
+            return (pos, T(kmer)), (i, kmer)
+        end
+        i += 1
+    end
+    return nothing
+end
+
+function Base.iterate(it::SpacedKmerIterator{T,S}, state::Tuple{Int,UInt64}) where {T,S<:BioSequence{<:FourBitNucs}}
+    i, kmer = state
+    filled = it.filled
+    i += it.increment
+    while i ≤ it.stop
+        nt = reinterpret(Int8, inbounds_getindex(it.seq, i))
+        if nt == 0xff # ambiguous
+            filled = 0
+        else
+            filled += 1
+            @inbounds val = kmerbits[nt + 1]
+            kmer = kmer << 2 | val
+        end
+        if filled == kmersize(T)
+            pos = i - kmersize(T) + 1
+            return (pos, T(kmer)), (i, kmer)
+        end
+        i += 1
+    end
+    return nothing
 end