--- a/src/minhash.jl
+++ b/src/minhash.jl
@@ -5,7 +5,6 @@
 #
 # This file is a part of BioJulia.
 # License is MIT: https://github.com/BioJulia/BioSequences.jl/blob/master/LICENSE.md
-using DataStructures: SortedSet
 
 """
 MinHash Sketch type