--- a/test/kmers/mismatches.jl
+++ b/test/kmers/mismatches.jl
@@ -7,7 +7,7 @@
         @test mismatches(a, b) === mismatches(b, a) === count
     end
 
-    for len in 0:32, _ in 1:10
+    for len in 1:32, _ in 1:10
         a = random_dna_kmer(len)
         b = random_dna_kmer(len)
         test_mismatches(DNAKmer(a), DNAKmer(b))