… CHANGELOG.md
… test/kmers/conversion.jl
… test/kmers/eachkmer.jl
… test/kmers/print.jl
… test/kmers/shuffle.jl
… test/kmers/transformations.jl
… test/runtests.jl