# Type specialization for faster kmer iteration and misc changes

## Types of changes

This PR implements the following changes:
_(Please tick any or all of the following that are applicable)_

* [:heavy_check_mark: ] :sparkles: New feature (A non-breaking change which adds functionality).
* [ :heavy_check_mark:] :bug: Bug fix (A non-breaking change, which fixes an issue).
* [ :heavy_check_mark:] :boom: Breaking change (fix or feature that would cause existing functionality to change).

## :clipboard: Additional detail

This pull requests has a variety of changes:

### Improved Kmer iteration speed

As discussed in [Issue 46](https://github.com/BioJulia/BioSequences.jl/issues/46), the previous implementation of `iterate` on `EachKmerIterator` was general, and did not specialize for the special cases of two-bit nucletide sequences, nor for cases where the iterator's `step` was 1.

Here, I have added three new types:
`AbstractKmerIterator`
`EveryKmerIterator <: AbstractKmerIterator`, for step == 1
`SpacedKmerIterator <: AbstractKmerIterator`, for step != 1

and removed the old `EachKmerIterator` type. Further, I have added a `length` method for 2-bit nucleotide kmer iterators, since their length can be known beforehand.


**Note**: It might be useful for the `each` function to take start and stop keywords, I have not added them in order to not change the interface.

**Justification**
Many bioinformatics workflows rely on kmers, precisely because they are fast. By changing the iteration protocol to improve speeds, many current and potential downstream applications of kmers such as kmer composition analysis or kmer distribution will be significantly sped up.

**Benchmark**
The performance difference depends on several factors: Sequence type, value of K, value of `step` and sequence length. 

In general, the speed relative to the current implementation is 4.5x in the typical case, 1.2x in the worst case and 27x in the best case.

In order to benchmark, I've used this simple function:
```
function bm(x)
    n = 0
    for i in x
        n += 1
    end
    return n
end
```

With the following results:
Before:
```
Len: 100    Bit: 2  K: 4    Step: 1    772.761 ns (0 allocations: 0 bytes)
Len: 100    Bit: 2  K: 4    Step: 4    323.026 ns (0 allocations: 0 bytes)
Len: 100    Bit: 2  K: 30   Step: 1    639.707 ns (0 allocations: 0 bytes)
Len: 100    Bit: 2  K: 30   Step: 4    317.760 ns (0 allocations: 0 bytes)
Len: 100    Bit: 4  K: 4    Step: 1    659.629 ns (0 allocations: 0 bytes)
Len: 100    Bit: 4  K: 4    Step: 4    280.278 ns (0 allocations: 0 bytes)
Len: 100    Bit: 4  K: 30   Step: 1    607.361 ns (0 allocations: 0 bytes)
Len: 100    Bit: 4  K: 30   Step: 4    276.466 ns (0 allocations: 0 bytes)
Len: 100000 Bit: 2  K: 4    Step: 1    776.198 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 2  K: 4    Step: 4    313.785 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 2  K: 4    Step: 100   12.514 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 2  K: 30   Step: 1    781.717 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 2  K: 30   Step: 4    349.243 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 2  K: 30   Step: 100   62.372 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 4  K: 4    Step: 1    659.685 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 4  K: 4    Step: 4    266.367 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 4  K: 4    Step: 100   10.692 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 4  K: 30   Step: 1    759.375 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 4  K: 30   Step: 4    310.910 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 4  K: 30   Step: 100   50.230 μs (0 allocations: 0 bytes)
```
After
```
Len: 100    Bit: 2  K: 4    Step: 1    57.360 ns (0 allocations: 0 bytes)
Len: 100    Bit: 2  K: 4    Step: 4    62.466 ns (0 allocations: 0 bytes)
Len: 100    Bit: 2  K: 30   Step: 1    55.463 ns (0 allocations: 0 bytes)
Len: 100    Bit: 2  K: 30   Step: 4    70.036 ns (0 allocations: 0 bytes)
Len: 100    Bit: 4  K: 4    Step: 1   151.853 ns (0 allocations: 0 bytes)
Len: 100    Bit: 4  K: 4    Step: 4   131.818 ns (0 allocations: 0 bytes)
Len: 100    Bit: 4  K: 30   Step: 1   148.653 ns (0 allocations: 0 bytes)
Len: 100    Bit: 4  K: 30   Step: 4   123.503 ns (0 allocations: 0 bytes)
Len: 100000 Bit: 2  K: 4    Step: 1    28.718 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 2  K: 4    Step: 4    35.829 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 2  K: 4    Step: 100   1.460 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 2  K: 30   Step: 1    28.694 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 2  K: 30   Step: 4    42.997 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 2  K: 30   Step: 100  18.374 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 4  K: 4    Step: 1   143.237 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 4  K: 4    Step: 4   121.794 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 4  K: 4    Step: 100   4.979 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 4  K: 30   Step: 1   143.241 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 4  K: 30   Step: 4   121.809 μs (0 allocations: 0 bytes)
Len: 100000 Bit: 4  K: 30   Step: 100  35.020 μs (0 allocations: 0 bytes)
```

### Disallow instantiation of zero-mers and negative kmers
__This change is breaking__

Previously, kmers with K of 0 or negative numbers could be instantiated. With these changes, attempting to do so throws an ArgumentError.
The rationale is both that zero or negative kmers make no sense biologically and thus should not be used, and also that they present as edge cases which complicates the code and leads to unexpected errors.

### Remove unused dependency DataStructures
Previously, DataStructures was listed as a dependency and imported in `minhash.jl` but not actually used in any code. I have simply removed DataStructures from `REQUIRE` and `minhash.jl`. This has no effect on the code.

### Change documentation and testing for reverse complementation
On the existing master branch, the documentation for `reverse_complements` states it leaves ambiguous nucleotides as-is, while in actuality, the existing code handles ambiguous nucleotides correctly. Also, the correct complementation and reverse-complementation of ambiguous nucleotides other than N are not tested for. I have amended the documentation and added test cases for ambiguous nucleotides.

### Minor speed increase in Composition for K > 8
Kmer Composition now increments its Dict more effectively resulting in a small speed increase. Other than the modest increase in speed, this is an inconsequential single-line change.


## :ballot_box_with_check: Checklist

- [:heavy_check_mark: ] :art: The changes implemented is consistent with the [julia style guide](https://docs.julialang.org/en/stable/manual/style-guide/).
- [ :heavy_check_mark:] :blue_book: I have updated and added relevant docstrings, in a manner consistent with the [documentation styleguide](https://docs.julialang.org/en/stable/manual/documentation/).
- [:heavy_check_mark: ] :blue_book: I have added or updated relevant user and developer manuals/documentation in `docs/src/`.
- [:heavy_check_mark: ] :ok: There are unit tests that cover the code changes I have made.
- [:heavy_check_mark: ] :ok: The unit tests cover my code changes AND they pass.
- [:heavy_check_mark: ] :pencil: I have added an entry to the `[UNRELEASED]` section of the manually curated `CHANGELOG.md` file for this repository.
- [:heavy_check_mark: ] :ok: All changes should be compatible with the latest stable version of Julia.
- [:heavy_check_mark: ] :thought_balloon: I have commented liberally for any complex pieces of internal code.