@jakobnissen 2018-10-15
> Are you finished with this @jakobnissen, or do you still have a bit more to do?
I just discovered that iterating over `eachkmer(seq, 4)` is ~700x times slower than `each(DNAKmer{4}, seq)` because the former is unable to type-specialize due to 4 being a value, not a type. It seems bad that such a performance trap awaits users, who will have little change to realize this subtle difference. What do you say we do? Rewrite `eachkmer` to take `Val{k} where {k}` or remove the `eachkmer` interface?