@BenJWard 2018-10-15
I would vote for removing the `eachkmer` interface. But that would mean changing the API which would require a version change. I'm doing API changes over on my fork too, so a v2.x of BioSequences is in the works. It depends on if you want to merge now with `eachkmer` (which would be a minor version bump with speed improvements) and I can just remove it later as part of my more API altering mods.