@jakobnissen 2018-10-15
Alright. I say leave it and merge it now, then. Note that with my changes, `eachkmer` is 4x slower than the existing `eachkmer` method, because it's even less certain about types than before now that there are multiple subtypes of `AbstractKmerIterator` it could return. However, with or without this PR merged, `eachkmer` is totally unsuitable for speed-sensitive code anyway, so in practice it doesn't change anything. Merge it. :)