@BenJWard 2018-10-15
I've just added warnings that should notfiy people that eachmer is slow and is being deprecated to discourage its use. I'll just make sure it doesen't screw up the tests and I'm happy to merge it.