@BenJWard 2018-10-15
Merged, thank you so much @jakobnissen for taking the time to do this!