@kescobo 2018-10-17
Nice work! Sorry about the DataStructures dep, that was my bad. At one point I was trying to use some special array type that I thought would be faster. Forgot to remove it when we decided it was the wrong path. Good catch!